+++ scripts/setup/create_and_fill_localPackages.ps1
@@ -1,0 +1,2 @@
+New-Item -ItemType Directory -Force -Path D:\Projects\IIChatTools\LocalPackages
+dotnet restore IIChatTools.sln --configfile NuGet.Config.online --packages D:\Projects\IIChatTools\LocalPackages --force